small refactor and merging with main
